Split into modeltypes for different model params and utils, and spectrum for spectrum specific matters
